requirements.txt
beautifulsoup4
fastapi
gunicorn
newspaper3k
numpy
Pillow
pydantic
python-dotenv
requests
uvicorn
